requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pymongo==4.8.0
motor==3.5.1
supabase==2.3.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
pydantic==2.5.3
